requirements.txt
dj-database-url
django
gunicorn
matplotlib
numpy
opencv-python
pillow
psycopg[binary]
tensorflow
torch
transformers
whitenoise
